flatten: textEditor
