--- packages/client/lib/app/core/widgets/text_editor/stack/presentation/presentation.dart
@@ -1,2 +1,0 @@
-export 'mobx/text_editor_store.dart';
-export "widget/text_editor.dart";
