--- packages/client/lib/app/core/widgets/text_editor/stack/stack.dart
@@ -1,1 +1,0 @@
-export "presentation/presentation.dart";
