+++ packages/client/lib/app/core/widgets/text_editor/text_editor_store.dart
